Moved graph options to dedicated file.  Removed mockup data from Network component.
