+++ client/assets/mockups/networkData.js
@@ -1,0 +1,97 @@
+//mockup data
+export const nodeList = [
+  {
+    type: 'ingress',
+    name: 'https://foo.example.com',
+  },
+  {
+    type: 'ingress',
+    name: 'https://bar.example.com',
+  },
+  {
+    type: 'service',
+    name: 'cluster_IP',
+    IP: '10.1.1.7',
+    port: 8080,
+    targetPort: 8080,
+  },
+  {
+    type: 'service',
+    name: 'cluster_IP',
+    port: 8080,
+    targetPort: 8080,
+  },
+  {
+    type: 'service',
+    name: 'cluster_IP',
+    port: 8080,
+    targetPort: 8080,
+  },
+  {
+    type: 'pod',
+    status: 'OK',
+    name: 'POST_1',
+    containers: 'post',
+    hostIP: '10.1.1.7',
+    podIP: '192.168.1.2',
+    volume: 'VOL_1',
+  },
+  {
+    type: 'pod',
+    status: 'ERROR',
+    name: 'POST_2',
+    containers: 'post',
+    hostIP: '10.1.1.8',
+    podIP: '192.168.1.2',
+    volume: 'VOL_1',
+  },
+  {
+    type: 'pod',
+    status: 'OK',
+    name: 'POST_3',
+    containers: 'post',
+    hostIP: '10.1.1.7',
+    podIP: '192.168.1.3',
+    volume: 'VOL_1',
+  },
+  {
+    type: 'pod',
+    status: 'OK',
+    name: 'COMMENT_1',
+    containers: 'comment',
+    hostIP: '10.1.1.8',
+    podIP: '192.168.1.3',
+    volume: 'VOL_2',
+  },
+  {
+    type: 'pod',
+    status: 'OK',
+    name: 'COMMENT_2',
+    containers: 'comment',
+    hostIP: '10.1.1.8',
+    podIP: '192.168.1.4',
+    volume: 'VOL_2',
+  },
+  {
+    type: 'pod',
+    status: 'OK',
+    name: 'MODERATION_1',
+    containers: 'moderation',
+    hostIP: '10.1.1.9',
+    podIP: '192.168.1.4',
+    volume: 'VOL_2',
+  },
+];
+
+//mockup data
+export const edgeList = [
+  { from: 0, to: 2 },
+  { from: 0, to: 3 },
+  { from: 1, to: 4 },
+  { from: 2, to: 5 },
+  { from: 2, to: 6 },
+  { from: 2, to: 7 },
+  { from: 3, to: 8 },
+  { from: 3, to: 9 },
+  { from: 4, to: 10 },
+];
